refactor(wave4-step2): tighten facade thinness for lockfile io and cache put
