+++ docs/contributing/SPLIT-REVIEW-PACK-wave4-step2.md
@@ -25,6 +25,9 @@ Executed and passing:
 - `cargo check -p assay-registry`
 - Lockfile/cache anchor subset tests
 - Delegation gates (facade -> `*_next` impl paths)
+- Facade-thinness tightening:
+  - `Lockfile::load/save` now delegate to `lockfile_next/io.rs`
+  - `PackCache::put` now delegates to `cache_next/put.rs`
 - Single-source gates:
   - ordering path only in `lockfile_next/format.rs`
   - atomic write/rename path only in `cache_next/io.rs`
@@ -34,3 +37,6 @@ No behavior/perf changes intended:
 - no output/error-string rewrites
 - no dependency/Cargo changes
 - no demo/ changes
+
+Atomic write note:
+- rename-based atomic write semantics are unchanged in Step2 (no fsync hardening added here).
