+++ scripts/ci/review-wave4-step2.sh
@@ -55,6 +55,8 @@ do
 done
 
 echo "== Wave4 Step2 delegation gates =="
+check_has_match 'lockfile_next::io::load_impl' crates/assay-registry/src/lockfile.rs
+check_has_match 'lockfile_next::io::save_impl' crates/assay-registry/src/lockfile.rs
 check_has_match 'lockfile_next::parse::parse_lockfile_impl' crates/assay-registry/src/lockfile.rs
 check_has_match 'lockfile_next::format::to_yaml_impl' crates/assay-registry/src/lockfile.rs
 check_has_match 'lockfile_next::format::add_pack_impl' crates/assay-registry/src/lockfile.rs
@@ -63,14 +65,22 @@ check_has_match 'lockfile_next::digest::verify_lockfile_impl' crates/assay-regis
 check_has_match 'lockfile_next::digest::check_lockfile_impl' crates/assay-registry/src/lockfile.rs
 check_has_match 'lockfile_next::digest::update_lockfile_impl' crates/assay-registry/src/lockfile.rs
 
+check_has_match 'cache_next::put::put_impl' crates/assay-registry/src/cache.rs
 check_has_match 'cache_next::keys::pack_dir_impl' crates/assay-registry/src/cache.rs
 check_has_match 'cache_next::io::default_cache_dir_impl' crates/assay-registry/src/cache.rs
 check_has_match 'cache_next::policy::parse_cache_control_expiry_impl' crates/assay-registry/src/cache.rs
 check_has_match 'cache_next::integrity::parse_signature_impl' crates/assay-registry/src/cache.rs
 check_has_match 'cache_next::io::write_atomic_impl' crates/assay-registry/src/cache.rs
 
+# Lockfile facade should no longer own direct fs/logging paths for load/save.
+if "$rg_bin" -n 'tokio::fs|tracing::info|fs::read_to_string|fs::write' crates/assay-registry/src/lockfile.rs; then
+  echo "lockfile facade still contains direct IO/logging ownership"
+  exit 1
+fi
+
 echo "== Wave4 Step2 single-source gates =="
 check_no_match_in_dir_excluding 'fs::rename\(' crates/assay-registry/src/cache_next io.rs
+check_no_match_in_dir_excluding 'create_dir_all\(' crates/assay-registry/src/cache_next put.rs
 check_no_match_in_dir_excluding 'max-age=' crates/assay-registry/src/cache_next policy.rs
 check_no_match_in_dir_excluding 'sort_by\(' crates/assay-registry/src/lockfile_next format.rs
